refactor(authz): enforce the procedure-level authz contract at server startup
Move servedServices / servedProcedures out of the _test file into a non-test
authz_served.go and add AssertRegistryMatchesServedProcedures, which server.start
now runs right after the service-level assertServedServicesMatch. A method added
to an already-mounted service is invisible to the service-level check; it now
fails startup fast instead of relying solely on AuthzInterceptor's runtime
fail-closed.

The pure core (assertRegistryCoversProcedures) is split out so both rejection
paths — a served procedure with no registry entry, and a stale entry — are
unit-tested with synthetic inputs. CI and boot read the same servedProcedures,
so the two cannot drift.
